API

Models

NearestNeighborModels.KNNClassifier — Type

KNNClassifier(;kwargs...)

K-Nearest Neighbors classifier: predicts the class associated with a new point by taking a vote over the classes of the K-nearest points.

Keywords Parameters

K::Int=5 : number of neighbors

algorithm::Symbol = :kdtree : one of (:kdtree, :brutetree, :balltree)

metric::Metric = Euclidean() : any Metric from Distances.jl for the distance between points. For algorithm = :kdtree only metrics which are of type Union{Distances.Chebyshev, Distances.Cityblock, Distances.Euclidean, Distances.Minkowski, Distances.WeightedCityblock, Distances.WeightedEuclidean, Distances.WeightedMinkowski} are supported.

leafsize::Int = algorithm == 10 : determines the number of points at which to stop splitting the tree. This option is ignored and always taken as 0 for algorithm = :brutetree, since brutetree isn't actually a tree.

reorder::Bool = true : if true then points which are close in distance are placed close in memory. In this case, a copy of the original data will be made so that the original data is left unmodified. Setting this to true can significantly improve performance of the specified algorithm (except :brutetree). This option is ignored and always taken as false for algorithm = :brutetree.

weights::KNNKernel=Uniform() : kernel used in assigning weights to the k-nearest neighbors for each observation. An instance of one of the types in list_kernels(). User-defined weighting functions can be passed by wrapping the function in a UDF kernel. If sample weights w are passed during machine construction e.g machine(model, X, y, w) then the weight assigned to each neighbor is the product of the KNNKernel generated weight and the corresponding neighbor sample weight.

See also the package documentation. For more information about the kernels see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

NearestNeighborModels.KNNRegressor — Type

KNNRegressor(;kwargs...)

K-Nearest Neighbors regressor: predicts the response associated with a new point by taking an weighted average of the response of the K-nearest points.

Keywords Parameters

K::Int=5 : number of neighbors

algorithm::Symbol = :kdtree : one of (:kdtree, :brutetree, :balltree)

metric::Metric = Euclidean() : any Metric from Distances.jl for the distance between points. For algorithm = :kdtree only metrics which are of type Union{Distances.Chebyshev, Distances.Cityblock, Distances.Euclidean, Distances.Minkowski, Distances.WeightedCityblock, Distances.WeightedEuclidean, Distances.WeightedMinkowski} are supported.

leafsize::Int = algorithm == 10 : determines the number of points at which to stop splitting the tree. This option is ignored and always taken as 0 for algorithm = :brutetree, since brutetree isn't actually a tree.

reorder::Bool = true : if true then points which are close in distance are placed close in memory. In this case, a copy of the original data will be made so that the original data is left unmodified. Setting this to true can significantly improve performance of the specified algorithm (except :brutetree). This option is ignored and always taken as false for algorithm = :brutetree.

weights::KNNKernel=Uniform() : kernel used in assigning weights to the k-nearest neighbors for each observation. An instance of one of the types in list_kernels(). User-defined weighting functions can be passed by wrapping the function in a UDF kernel. If sample weights w are passed during machine construction e.g machine(model, X, y, w) then the weight assigned to each neighbor is the product of the KNNKernel generated weight and the corresponding neighbor sample weight.

See also the package documentation. For more information about the kernels see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

NearestNeighborModels.MultitargetKNNRegressor — Type

MultitargetKNNRegressor(;kwargs...)

K-Nearest Neighbors regressor: predicts the response associated with a new point by taking an weighted average of the response of the K-nearest points.

Keywords Parameters

K::Int=5 : number of neighbors

algorithm::Symbol = :kdtree : one of (:kdtree, :brutetree, :balltree)

metric::Metric = Euclidean() : any Metric from Distances.jl for the distance between points. For algorithm = :kdtree only metrics which are of type Union{Distances.Chebyshev, Distances.Cityblock, Distances.Euclidean, Distances.Minkowski, Distances.WeightedCityblock, Distances.WeightedEuclidean, Distances.WeightedMinkowski} are supported.

leafsize::Int = algorithm == 10 : determines the number of points at which to stop splitting the tree. This option is ignored and always taken as 0 for algorithm = :brutetree, since brutetree isn't actually a tree.

reorder::Bool = true : if true then points which are close in distance are placed close in memory. In this case, a copy of the original data will be made so that the original data is left unmodified. Setting this to true can significantly improve performance of the specified algorithm (except :brutetree). This option is ignored and always taken as false for algorithm = :brutetree.

weights::KNNKernel=Uniform() : kernel used in assigning weights to the k-nearest neighbors for each observation. An instance of one of the types in list_kernels(). User-defined weighting functions can be passed by wrapping the function in a UDF kernel. If sample weights w are passed during machine construction e.g machine(model, X, y, w) then the weight assigned to each neighbor is the product of the KNNKernel generated weight and the corresponding neighbor sample weight.

See also the package documentation. For more information about the kernels see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

NearestNeighborModels.MultitargetKNNClassifier — Type

MultitargetKNNClassifier(;kwargs...)

K-Nearest Neighbors classifier: predicts the class associated with a new point by taking a vote over the classes of the K-nearest points.

Keywords Parameters

K::Int=5 : number of neighbors

algorithm::Symbol = :kdtree : one of (:kdtree, :brutetree, :balltree)

metric::Metric = Euclidean() : any Metric from Distances.jl for the distance between points. For algorithm = :kdtree only metrics which are of type Union{Distances.Chebyshev, Distances.Cityblock, Distances.Euclidean, Distances.Minkowski, Distances.WeightedCityblock, Distances.WeightedEuclidean, Distances.WeightedMinkowski} are supported.

leafsize::Int = algorithm == 10 : determines the number of points at which to stop splitting the tree. This option is ignored and always taken as 0 for algorithm = :brutetree, since brutetree isn't actually a tree.

reorder::Bool = true : if true then points which are close in distance are placed close in memory. In this case, a copy of the original data will be made so that the original data is left unmodified. Setting this to true can significantly improve performance of the specified algorithm (except :brutetree). This option is ignored and always taken as false for algorithm = :brutetree.

weights::KNNKernel=Uniform() : kernel used in assigning weights to the k-nearest neighbors for each observation. An instance of one of the types in list_kernels(). User-defined weighting functions can be passed by wrapping the function in a UDF kernel. If sample weights w are passed during machine construction e.g machine(model, X, y, w) then the weight assigned to each neighbor is the product of the KNNKernel generated weight and the corresponding neighbor sample weight.

output_type::Type{<:MultiUnivariateFinite}=DictTable : One of (ColumnTable, DictTable). The type of table type to use for predictions. Setting to ColumnTable might improve performance for narrow tables while setting to DictTable improves performance for wide tables.

See also the package documentation. For more information about the kernels see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

Kernels

NearestNeighborModels.KNNKernel — Type

KNNKernel

Abstract super type for all weighting kernels

NearestNeighborModels.list_kernels — Function

list_kernels()

Lists all implemented KNN weighting kernels

NearestNeighborModels.DualU — Type

DualU()

Assigns the closest neighbor a weight of 1, the furthest neighbor weight 0 and the others are scaled between by a mapping.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: DualD

NearestNeighborModels.DualD — Type

DualD()

Assigns the closest neighbor a weight of 1, the furthest neighbor weight 0 and the others are scaled between by a mapping.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: DualU

NearestNeighborModels.Dudani — Type

Dudani()

Assigns the closest neighbor a weight of 1, the furthest neighbor weight 0 and the others are scaled between by a linear mapping.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: Macleod, DualU, DualD

NearestNeighborModels.Fibonacci — Type

Fibonacci()

Assigns neighbors weights corresponding to fibonacci numbers starting from the furthest neighbor. i.e the furthest neighbor a weight of 1, the second furthest neighbor a weight of 1 and the third furthest neighbor a weight of 2 and so on.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

NearestNeighborModels.Inverse — Type

Inverse()

Assigns each neighbor a weight equal to the inverse of the corresponsting distance of the neighbor.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: ISquared

NearestNeighborModels.ISquared — Type

ISquared()

Assigns each neighbor a weight equal to the inverse of the corresponsting squared-distance of the neighbor.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

NearestNeighborModels.Macleod — Type

Macleod(;a::Real= 0.0)

Assigns the closest neighbor a weight of 1, the furthest neighbor weight 0 and the others are scaled between by a linear mapping.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: Dudani, DualU, DualD

NearestNeighborModels.Rank — Type

Rank()

Assigns each neighbor a weight as a rank such that the closest neighbor get's a weight of 1 and the Kth closest neighbor gets a weight of K.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: ReciprocalRank

NearestNeighborModels.ReciprocalRank — Type

ReciprocalRank(;a::Real= 0.0)

Assigns each closest neighbor a weight which is equal to the reciprocal of it's rank. i.e the closest neighbor get's a weight of 1 and the Kth closest weight get's a weight of 1/K

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.

see also: Rank

NearestNeighborModels.UDK — Type

UDK

Alias for UserDefinedKernel

NearestNeighborModels.Uniform — Type

Uniform()

assigns uniform weights to all k-nearest neighbors.

see also: Inverse, ISquared, Zavreal

NearestNeighborModels.UserDefinedKernel — Type

UserDefinedKernel(;func::Function = x->nothing, sort::Bool=false)

Wrap a user defined nearest neighbors weighting function func as a KNNKernel.

Keywords

func : user-defined nearest neighbors weighting function. The function should have the signature func(dists_matrix)::Union{Nothing, <:AbstractMatrix}. The dists_matrix is a n by K nearest neighbors distances matrix where n is the number of samples in the test dataset and K is number of neighbors. func should either output nothing or an AbstractMatrix of the same shape as dists_matrix. If func(dists_matrix) returns nothing then all k-nearest neighbors in each row are assign equal weights.

sort : if true requests that the dists_matrix be sorted before being passed to func. The sort is done in a manner that puts the k-nearest neighbors in each row of dists_matrix in acesending order .

NearestNeighborModels.Zavreal — Type

Zavreal(;s::Real = 0.0, a::Real=1.0)

Assigns each neighbor an exponential weight given by $e^{ - α ⋅ d_i^{eta}}$ where α and β are constants and dᵢ is the distance of the given neighbor.

For more information see the paper by Geler et.al Comparison of different weighting schemes for the kNN classifier on time-series data.